Refactor pdfs
Refactors the pdf plotting code with all the changes from the timeseries/1-panels. Operations are modularized, inputs are reworked, icecream is added, filenames are changed to match 1-panel scheme. Additionally, changes from pdfs are made to other code--most notably, the spatial averaging is added as an input to manage_area in process_glens_fun (areaAvgBool, defaults to true). Tidying comes next, but this adds the functionality!