Implement plot_difference_globes
Difference globes are now fully functionized! Instead of the previous script-based interface, the difference globes can now be called individually using a dictionary-based input system. Example dictionaries are given in the comment block at the start of plot_difference_globes. This version of run_test_script includes an example of interacting with the four functions. open_data is added to process_glens_fun to facilitate easy access within the functions.