Handle ensemble means
Adds code to more intelligently handle figures with ensemble means. Number of ensemble members available in the means is added to the titles or legends of figures, as appropriate. Note that these numbers are the MAX AVAILABLE MEMBERS, and are not detected from the filenames. Additionally, employs plt.close() to improve memory handling for large numbers of figure generation.