Add better vmin-vmax for strat baseline diff globe
Stratosphere sum baseline difference globe has been blowing out the colorbars on almost every image due to the asymmetry between the negative 99th quantile and the 1st quantile. This introduces a method to compare the two quantiles and use whichever has the widest range. In the interest of simplicity, I'm only adding this to the panel that's been a problem, but the same method could be used elsewhere if needed.