Interface GLENS2 with pdfs and ensemble spaghetti
Ensemble spaghetti timeseries works with GLENS2 globally, but not for area plots yet.
PDFs work with GLENS2 (with the valid period check disabled). Things that were already cumbersome (e.g. generate_labels) are no better than they were before. Automatic color generation for GEO2-4.5 is fully implemented with a set of blues analogous to the set of GEO8.5 purples. GEO2-4.5 colors use the same luminance values as the purples and reds for GEO8.5 and RCP8.5, respectively.