Modularize panel-making in difference globes
Data processing for panels is now handled in a function called make_panels in fun_plot tools. Refactor (almost) all difference globes to use this. Single-scenario difference globes will need a little effort to work with this function because of the handling of the extra period of interest. Vertical difference globes are commented and moved to end of the function.